+++ federation/entities/diaspora/utils.py
@@ -1,4 +1,6 @@
 # -*- coding: utf-8 -*-
+import inspect
+
 from dateutil.tz import tzlocal, tzutc
 from lxml import etree
 
@@ -32,3 +34,17 @@ def struct_to_xml(node, struct):
     for obj in struct:
         for k, v in obj.items():
             etree.SubElement(node, k).text = v
+
+
+def get_base_attributes(entity):
+    """Build a dict of attributes of an entity.
+
+    Returns attributes and their values, ignoring any properties, functions and anything that starts
+    with an underscore.
+    """
+    attributes = {}
+    cls = entity.__class__
+    for attr, _ in inspect.getmembers(cls, lambda o: not isinstance(o, property) and not inspect.isroutine(o)):
+        if not attr.startswith("_"):
+            attributes[attr] = getattr(entity, attr)
+    return attributes